requirements.txt
pandas
scikit-learn
joblib
openai
python-dotenv
